requirements.txt
opencv_python==4.1.1.26
matplotlib==3.1.1
numpy==1.17.0
torchvision==0.4.0a0+6b959ee
torch==1.3.0
pdf2image==1.6.0
Pillow==6.2.1
